Run the full functional suite on GitHub under the known-failures xfail model
Replaces the allowlist with the xfail model on the GitHub lane, so the workflow
becomes a source-built twin of the OSS lane instead of a curated subset.

**No more allowlist.** `config/allowlist.yml` (10,502 lines) and
`tools/conftest_allowlist.py` are deleted. The full ~50k-test suite now runs
under `tools/conftest_known_failures.py`: a per-gateway failing list is
`xfail(strict=True)`, a flaky list is `xfail(strict=False)`, and the gate fails
on any residual failed/error or on an `XPASS(strict)` where a listed
expected-failure starts passing. An allowlist could only ever assert that a
hand-picked subset stays green; this asserts the whole suite's shape and makes
a newly-passing test a signal rather than silence.

**Engine-crasher skip list.** `config/ci_crash_tests.txt` (and the GitHub
subset `config/github_ci_crash_tests.txt`) hold tests that crash the backend.
They are applied as `skip`, not xfail: an xfailed crasher still runs, takes the
engine down, and cascades connection errors onto every concurrent test. The
list is engine-level rather than per-gateway and accepts file (`.py`) and
directory (`/`) prefixes so it survives suite pin bumps.

**Source build, split across matrix legs.** The workflow builds the engine and
gateway host once and ships the installed tree to the legs as a single tar
computed from a before/after path-set difference over the install roots (path
sets, not mtimes - `make install` preserves source timestamps often enough that
`find -newer` yields a silently thin tar whose legs then fail on a missing
header). Six parallel legs each run a deterministic stride slice with `-n auto`;
a `no_parallel` leg runs serially. `--expected-ids` makes green require a
recorded outcome for every assigned test, so lost xdist workers cannot silently
drop coverage.

**verify-splits fails closed on universe skew.** Each leg slices a universe it
collected itself, so the slices only tile the suite if every leg collected the
same one. A leg whose collection silently omits a node shifts every later
position by one and starts running the next leg's tests while its own share
runs on no leg - and every leg still reports PASS. The new job asserts all
parallel legs published an identical universe fingerprint.

**Build caching.** ccache for the C build (PATH masquerade, since only the
unprivileged `make` needs it), cargo tool and registry caches, and the rust
toolchain / cargo-pgrx setup moved to the background so it overlaps the engine
build instead of adding ~8 cold minutes in front of it.

**Baseline.** The OSS failing/flaky pair is regenerated for suite 5a50ef4 from
two full runs on this exact stack and cross-checked against the backend gate's
passing lists, with per-gateway divergences omitted rather than assumed.

Robustness fixes along the way: network fetches in the setup scripts retry and
the suite is shallow-fetched; `run-functional-tests.sh` preserves a nonzero
pytest exit so the gate cannot go falsely green; and the gate is confined to
this repository's paths.
