+++ documentdb-local/functional-tests/config/github_ci_crash_tests.txt
@@ -1,0 +1,33 @@
+# Engine-level crasher skips for the GITHUB functional gate (OSS stack on
+# ubuntu hosted runners). Same semantics as ci_crash_tests.txt — skip, never
+# xfail — but scoped to what is genuinely the ENGINE's fault, so it holds on
+# any runner:
+#   * the $redact SIGSEGV family (aggregate_cursor_first_page),
+#   * OSS-lane session wedgers proven on this gate by two-run fingerprints.
+#
+# What is deliberately NOT here: the ADO index-build-convoy sections. Those
+# are an ADO-AGENT property (2 background index-build executors convoying on
+# 4-cpu azl3 containers), not an engine defect, and ci_crash_tests.txt keeps
+# them for the ADO lanes. Measured on GitHub run 30667428483 with them
+# re-enabled: 2,564 tests ran, ZERO exceeded 60s, no worker deaths, and legs
+# grew only ~1-2 minutes — the convoy does not occur on these runners. Of the
+# re-enabled set 1,323 PASS (real coverage that would otherwise be invisible)
+# and the failures are carried as ordinary strict entries in
+# oss_ci_failing_tests.txt.
+#
+# Keep in sync with ci_crash_tests.txt for the engine-level sections only.
+
+docdb_functional_tests/documentdb_tests/compatibility/tests/core/indexes/commands/createIndexes/test_createIndexes_field_behavior.py::test_createIndexes_field_behavior[documentdb-bindata_values]
+docdb_functional_tests/documentdb_tests/compatibility/tests/core/operator/expressions/string/regexFindAll/test_regexFindAll_size_limit.py
+docdb_functional_tests/documentdb_tests/compatibility/tests/core/operator/expressions/string/replaceAll/test_replaceAll_size_limit.py
+docdb_functional_tests/documentdb_tests/compatibility/tests/core/operator/stages/merge/test_stages_merge_on_matching.py
+docdb_functional_tests/documentdb_tests/compatibility/tests/core/operator/stages/redact/
+
+# --- Forced here by the one-list rule, not by a crash on this runner --------
+# conftest_known_failures.py requires an EXACT node ID to live in exactly one
+# list, and oss_ci_failing_tests.txt is shared with the backend gate's OSS lane
+# — whose skip list already carries this exact ID as a queue-victim. Carrying it
+# as a strict entry there too would make the plugin raise on that lane and
+# collect zero tests, so it is skipped here instead. Restore it to
+# oss_ci_failing_tests.txt if that lane ever stops skipping it by exact ID.
+docdb_functional_tests/documentdb_tests/compatibility/tests/system/administration/commands/setIndexCommitQuorum/test_smoke_setIndexCommitQuorum.py::test_smoke_setIndexCommitQuorum[documentdb]
